Separate start, menu and actions from main.py
